# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Api
 * StruturaCheck - Structural Defect Detection System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Analysis {
  id: number;
  fileName: string;
  /** bridge, road, wall, building */
  structureType: string;
  /** high, medium, low, none */
  severity: string;
  /** pending, processing, completed, failed */
  status: string;
  /** @nullable */
  defectCount?: number | null;
  /** @nullable */
  confidenceScore?: number | null;
  /** @nullable */
  analysisSpeedMs?: number | null;
  /** @nullable */
  originalImageUrl?: string | null;
  /** @nullable */
  maskedImageUrl?: string | null;
  /**
     * JSON array of defect type strings
     * @nullable
     */
  defectTypes?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  buildingAge?: number | null;
  /** @nullable */
  numberOfFloors?: number | null;
  /** @nullable */
  materialType?: string | null;
  /** @nullable */
  healthScore?: number | null;
  createdAt: string;
}

export interface AnalysisInput {
  fileName: string;
  structureType: string;
  notes?: string;
}

export interface StatsSummary {
  totalImagesChecked: number;
  totalDefectsFound: number;
  avgAnalysisSpeedMs: number;
  highSeverityCount: number;
  mediumSeverityCount: number;
  lowSeverityCount: number;
  trendPercent: number;
}

export interface HistoryPoint {
  date: string;
  defectsFound: number;
  imagesChecked: number;
}

export interface DistributionItem {
  type: string;
  count: number;
  percentage: number;
}

export interface Recommendation {
  id: number;
  analysisId: number;
  /** critical, warning, safe */
  severity: string;
  title: string;
  description: string;
  /** Markdown content explaining the AI reasoning */
  reasoning: string;
  workOrderGenerated?: boolean;
  /** @nullable */
  reminderDate?: string | null;
  createdAt: string;
}

export interface WorkOrder {
  id: number;
  recommendationId: number;
  reference: string;
  createdAt: string;
}

export interface ReminderInput {
  reminderDate: string;
}

export interface Report {
  analysisId: number;
  fileName: string;
  structureType: string;
  severity: string;
  createdAt: string;
  defectCount: number;
  /** @nullable */
  confidenceScore?: number | null;
  /** @nullable */
  buildingAge?: number | null;
  /** @nullable */
  numberOfFloors?: number | null;
  /** @nullable */
  materialType?: string | null;
  /** @nullable */
  healthScore?: number | null;
  recommendations: Recommendation[];
}