# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StruturaCheck - Structural Defect Detection System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List all analyses
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"

    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Create a new analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a single analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found

    delete:
      operationId: deleteAnalysis
      tags: [analyses]
      summary: Delete an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reports/{analysisId}:
    get:
      operationId: downloadReport
      tags: [analyses]
      summary: Get report data for an analysis
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
components:
  schemas:
...
    Analysis:
      type: object
      required: [id, fileName, structureType, severity, status, createdAt]
      properties:
        id:
          type: integer
        fileName:
          type: string
        structureType:
          type: string
          description: "bridge, road, wall, building"
        severity:
          type: string
          description: "high, medium, low, none"
        status:
          type: string
          description: "pending, processing, completed, failed"
        defectCount:
          type: ["integer", "null"]
        confidenceScore:
          type: ["number", "null"]
        analysisSpeedMs:
          type: ["integer", "null"]
        originalImageUrl:
          type: ["string", "null"]
        maskedImageUrl:
          type: ["string", "null"]
        defectTypes:
          type: ["string", "null"]
          description: "JSON array of defect type strings"
        notes:
          type: ["string", "null"]
        buildingAge:
          type: ["integer", "null"]
        numberOfFloors:
          type: ["integer", "null"]
        materialType:
          type: ["string", "null"]
        healthScore:
          type: ["integer", "null"]
        createdAt:
          type: string
    AnalysisInput:
      type: object
      required: [fileName, structureType]
      properties:
        fileName:
          type: string
        structureType:
          type: string
        notes:
          type: string
...
    Recommendation:
      type: object
      required: [id, analysisId, severity, title, description, reasoning, createdAt]
      properties:
        id:
          type: integer
        analysisId:
          type: integer
        severity:
          type: string
          description: "critical, warning, safe"
        title:
          type: string
        description:
          type: string
        reasoning:
          type: string
          description: "Markdown content explaining the AI reasoning"
        workOrderGenerated:
          type: boolean
        reminderDate:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Report:
      type: object
      required: [analysisId, fileName, structureType, severity, createdAt, defectCount, recommendations]
      properties:
        analysisId:
          type: integer
        fileName:
          type: string
        structureType:
          type: string
        severity:
          type: string
        createdAt:
          type: string
        defectCount:
          type: integer
        confidenceScore:
          type: ["number", "null"]
        buildingAge:
          type: ["integer", "null"]
        numberOfFloors:
          type: ["integer", "null"]
        materialType:
          type: ["string", "null"]
        healthScore:
          type: ["integer", "null"]
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"