# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Api
 * StruturaCheck - Structural Defect Detection System API
 * OpenAPI spec version: 0.1.0
 */

export interface Analysis {
  id: number;
  fileName: string;
  /** bridge, road, wall, building */
  structureType: string;
  /** high, medium, low, none */
  severity: string;
  /** pending, processing, completed, failed */
  status: string;
  /** @nullable */
  defectCount?: number | null;
  /** @nullable */
  confidenceScore?: number | null;
  /** @nullable */
  analysisSpeedMs?: number | null;
  /** @nullable */
  originalImageUrl?: string | null;
  /** @nullable */
  maskedImageUrl?: string | null;
  /**
     * JSON array of defect type strings
     * @nullable
     */
  defectTypes?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  buildingAge?: number | null;
  /** @nullable */
  numberOfFloors?: number | null;
  /** @nullable */
  materialType?: string | null;
  /** @nullable */
  healthScore?: number | null;
  createdAt: string;
}